Fix saving employees.
Previously, trying to save an employee without reselecting the location and agency would result in a failure. This closes #28
