+++ app/Http/Requests/Employees/UpdateEmployeeRequest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Hourglass\Http\Requests;
+namespace Hourglass\Http\Requests\Employees;
 
 use Hourglass\Http\Requests\Request;
 use RedCrystal\ValidationRules\ExistsValidationRule;
 use RedCrystal\ValidationRules\UniqueValidationRule;
 
-class CreateEmployeeRequest extends Request
+class UpdateEmployeeRequest extends Request
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -30,19 +30,21 @@ class CreateEmployeeRequest extends Request
             'name' => 'required',
             'terminal_key' => (new UniqueValidationRule('employees', 'terminal_key'))
                 ->where('account_id', $accountId)
+                ->ignore($this->get('id'))
                 ->whereNull('deleted_at')
                 ->toString(),
-            'location_id' => [
+            'location.id' => [
                 'required',
                 (new ExistsValidationRule('locations', 'id'))
                     ->where('account_id', $accountId)
                     ->toString(),
             ],
-            'agency_id' => [
+            'agency.id' => [
+                'required',
                 (new ExistsValidationRule('agencies', 'id'))
                     ->where('account_id', $accountId)
-                    ->toString()
-            ]
+                    ->toString(),
+            ],
         ];
     }
 }
